refactor get user profile
