+++ expo-zustand-styled-components/src/services/queries/viewer-info.ts
@@ -1,0 +1,9 @@
+export const LOGGEDIN_USER_PROFILE = `
+  query LoggedinUser {
+    viewer {
+      avatarUrl
+      name
+      login
+    }
+  }
+`;
